@@ -24,11 +24,12 @@ public class DrawingView extends View {
     private double mBrushSize = 0.0;
     private int color = Color.BLACK;
     private Canvas canvas;
-    private ArrayList<CustomPath> mPaths;
+    private final ArrayList<CustomPath> mPaths;
 
 
-    public DrawingView(Context context, @Nullable AttributeSet attrs) {
+    public DrawingView(Context context, @Nullable AttributeSet attrs, ArrayList<CustomPath> mPaths) {
         super(context, attrs);
+        this.mPaths = mPaths;
     }
 
     private void setUpDrawing() {
@@ -85,9 +86,7 @@ public class DrawingView extends View {
 
         switch (event.getAction()) {
 
-            case 1:
-                int actionDown = MotionEvent.ACTION_DOWN;
-                 {
+            case MotionEvent.ACTION_DOWN -> {
                      int gtCl = mDrawPath.getColor();
                      gtCl = color;
                      float mDBrush = mDrawPath.getBrushThickness();
@@ -103,9 +102,7 @@ public class DrawingView extends View {
             }
             break;
 
-            case 2:
-                int actionMove = MotionEvent.ACTION_MOVE;
-            {
+            case MotionEvent.ACTION_MOVE -> {
                 if (touchX != null) {
                     if (touchY != null) {
                         mDrawPath.lineTo(touchX, touchY);
@@ -114,9 +111,7 @@ public class DrawingView extends View {
             }
             break;
 
-            case 3:
-                int actionUp = MotionEvent.ACTION_UP;
-            {
+            case MotionEvent.ACTION_UP -> {
                 mPaths.add(mDrawPath);
                 mDrawPath = new CustomPath(color, (float) mBrushSize);
             }
@@ -138,7 +133,6 @@ public class DrawingView extends View {
 
         strokeWidth = (float) mBrushSize;
     }
-
     protected void setColor(String newColor){
         color = Color.parseColor(newColor);
         int gtColor = mDrawPaint.getColor();
