@@ -4,10 +4,8 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.widget.TextView;
 
 import com.codelab.simplesignin.databinding.ActivityWelcomeBinding;
-import com.google.android.material.button.MaterialButton;
 
 public class WelcomeActivity extends AppCompatActivity {
     private ActivityWelcomeBinding mBinding;
@@ -18,8 +16,8 @@ public class WelcomeActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(mBinding.getRoot());
 
-        mBinding.logoutBtn.setOnClickListener(view -> {
-            startActivity(new Intent(this, MainActivity.class));
-        });
+        mBinding.logoutBtn.setOnClickListener(view -> startActivity(new Intent(this, MainActivity.class)));
+
+        mBinding.inputText.setText("Welcome");
     }
 }
